pyproject.toml
[project]
name = "ai-studio-story"
version = "1.0.0"
description = "A set of Python tool to help generate stories for Lunii device"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "Pillow>=10.0.0",
    "pydub>=0.25.0",
    "PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["generate_cover.py", "generate_audio.py"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
